Изменено хранение токенов доступа (теперь для каждого пользователя хранится свой токен)

+++ room-booking-app-backend/src/main/java/com/ppteam/roombookingapp/controllers/SubscriptionController.java
@@ -3,17 +3,15 @@ package com.ppteam.roombookingapp.controllers;
 import com.corundumstudio.socketio.AckRequest;
 import com.corundumstudio.socketio.SocketIOClient;
 import com.corundumstudio.socketio.SocketIOServer;
-import com.corundumstudio.socketio.listener.ConnectListener;
 import com.corundumstudio.socketio.listener.DataListener;
-import com.corundumstudio.socketio.listener.DisconnectListener;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import com.microsoft.graph.models.ChangeType;
 import com.microsoft.graph.models.Subscription;
 import com.microsoft.graph.requests.GraphServiceClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,34 +30,23 @@ public class SubscriptionController {
     @Autowired
     private SubscriptionStoreService subscriptionStoreService;
     @Autowired
-    private AccessTokenService accessTokenService;
-    static final String notificationHost = "https://8ba5-185-42-144-194.eu.ngrok.io";
+    private AccessTokenStoreService accessTokenStoreService;
+    static final String notificationHost = "https://d10c-185-42-144-194.eu.ngrok.io";
     static final int subscriptionLifetimeInMinutes = 1;
     static final int handicapInSeconds = 3;
     private final SocketIOServer socketIOServer;
 
     public SubscriptionController(SocketIOServer socketIOServer) {
         this.socketIOServer = socketIOServer;
-
-        this.socketIOServer.addConnectListener(new ConnectListener() {
-            @Override
-            public void onConnect(SocketIOClient socketIOClient) {
-                log.info("Клиент {} подключился", socketIOClient.getSessionId());
-            }
-        });
-        this.socketIOServer.addDisconnectListener(new DisconnectListener() {
-            @Override
-            public void onDisconnect(SocketIOClient socketIOClient) {
-                log.info("Клиент {} отключился", socketIOClient.getSessionId());
-            }
-        });
-
         this.socketIOServer.addEventListener("join_calendar_room", String.class, new DataListener<String>() {
             @Override
-            public void onData(SocketIOClient client, String calApiId, AckRequest ackRequest) throws Exception {
+            public void onData(SocketIOClient client, String request, AckRequest ackRequest) throws Exception {
+                JsonObject requestJson = JsonParser.parseString(request).getAsJsonObject();
+                String userId = requestJson.get("userId").getAsString().substring(19, 36).replace("-", "");
+                String calApiId = requestJson.get("calApiId").getAsString();
                 log.info("Клиент {} создал комнату для календаря {}", client.getSessionId(), calApiId);
                 client.joinRoom(calApiId);
-                createSubForCalendarIfNotExists(calApiId);
+                createSubForCalendarIfNotExists(calApiId, userId);
             }
         });
         this.socketIOServer.addEventListener("leave_calendar_room", String.class, new DataListener<String>() {
@@ -72,24 +59,28 @@ public class SubscriptionController {
     }
 
     /**
-     * Создает подписку для указанного ID календаря Outlook, если для данного ID не существует подписки.
+     * Создает подписку для указанного ID календаря Outlook и пользователя, если для данного ID календаря не существует
+     * подписки.
      *
      * @param calendarApiId ID календаря Outlook
+     * @param userId ID пользователя Outlook
      */
-    public void createSubForCalendarIfNotExists(String calendarApiId) {
-        if (subscriptionStoreService.hasActiveSubscriptionForCalendarId(calendarApiId)) {
+    public void createSubForCalendarIfNotExists(String calendarApiId, String userId) {
+        if (subscriptionStoreService.hasActiveSubscriptionForCalendarId(calendarApiId) ||
+                !accessTokenStoreService.hasAccessTokenForUserId(userId)) {
             return;
         }
-        GraphServiceClient graphClient = GraphClientHelper.getGraphClient(accessTokenService.getAccessToken());
+        GraphServiceClient graphClient = GraphClientHelper.getGraphClient(accessTokenStoreService.getAccessTokenByUserId(userId));
         Subscription subscriptionRequest = new Subscription();
         subscriptionRequest.changeType = ChangeType.CREATED + ", " + ChangeType.UPDATED + ", " + ChangeType.DELETED;
         subscriptionRequest.notificationUrl = notificationHost + "/listen";
-        subscriptionRequest.resource = "me/calendars/" + calendarApiId + "/events";
+        subscriptionRequest.resource = "Users/" + userId + "/calendars/" + calendarApiId + "/events";
         subscriptionRequest.expirationDateTime = OffsetDateTime.now().plusMinutes(subscriptionLifetimeInMinutes);
         CompletableFuture<Subscription> subscriptionFuture =
                 graphClient.subscriptions().buildRequest().postAsync(subscriptionRequest);
         subscriptionFuture.thenAccept(subscription -> {
-            subscriptionStoreService.addSubscription(subscription.id, subscription.resource, subscription.expirationDateTime);
+            subscriptionStoreService.addSubscription(subscription.id, subscription.resource,
+                    subscription.expirationDateTime, userId);
             log.info("Создана подписка: {} для ресурса: {}", subscription.id, subscription.resource);
         });
     }
@@ -119,13 +110,17 @@ public class SubscriptionController {
         log.info("Начато обновление подписок");
         List<SubscriptionRecord> subscriptions = subscriptionStoreService.getAllSubscriptions();
         if (!subscriptions.isEmpty()) {
-            GraphServiceClient graphClient = GraphClientHelper.getGraphClient(accessTokenService.getAccessToken());
             for (SubscriptionRecord subscription : subscriptions) {
-                String calendarApiId = SubscriptionStoreService.getCalendarApiIdFromResource(subscription.resource);
-                if (!socketIOServer.getRoomOperations(calendarApiId).getClients().isEmpty()) {
-                    updateSubscription(subscription.subscriptionId, graphClient);
-                } else {
-                    deleteSubscription(calendarApiId, graphClient);
+                String userId = subscription.resource.split("/")[1];
+                if (accessTokenStoreService.hasAccessTokenForUserId(userId)) {
+                    GraphServiceClient graphClient = GraphClientHelper
+                            .getGraphClient(accessTokenStoreService.getAccessTokenByUserId(userId));
+                    String calendarApiId = SubscriptionStoreService.getCalendarApiIdFromResource(subscription.resource);
+                    if (!socketIOServer.getRoomOperations(calendarApiId).getClients().isEmpty()) {
+                        updateSubscription(subscription.subscriptionId, graphClient);
+                    } else {
+                        deleteSubscription(calendarApiId, graphClient);
+                    }
                 }
             }
         }
@@ -139,9 +134,9 @@ public class SubscriptionController {
      */
     public void deleteSubscription(String calendarApiId, GraphServiceClient graphClient) {
         List<SubscriptionRecord> subscriptions = subscriptionStoreService.getAllSubscriptions();
-        String calendarAsResource = "me/calendars/" + calendarApiId + "/events";
         for (SubscriptionRecord subscription : subscriptions) {
-            if (Objects.equals(subscription.resource, calendarAsResource)) {
+            String subscriptionCalendarApiId = SubscriptionStoreService.getCalendarApiIdFromResource(subscription.resource);
+            if (Objects.equals(subscriptionCalendarApiId, calendarApiId)) {
                 graphClient.subscriptions(subscription.subscriptionId)
                         .buildRequest().deleteAsync().thenRun( () -> {
                             log.info("Удалена подписка: {}", subscription.subscriptionId);
